refactor: UI的な部品とロジック部分とを分割・整理。
