+++ typescript/src/checklist/domain/CheckListResult.ts
