+++ typescript/src/checklist/domain/CheckListResults.ts
