+++ typescript/src/checklist/domain/CheckResult.ts
